refactor(packageparser): move parsing the `exports` field into a separate module
